update dir structure to separate components
