+++ src/components/note/DeleteNote.js
@@ -2,7 +2,7 @@ import React, { Component, Fragment } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
 import PropTypes from 'prop-types';
 // components
-import AaButton from '../util/AaButton';
+import AaButton from '../../util/AaButton';
 // mui
 import Button from '@material-ui/core/Button';
 import Dialog from '@material-ui/core/Dialog';
@@ -12,7 +12,7 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DeleteOutline from '@material-ui/icons/DeleteOutline';
 
 import { connect } from 'react-redux';
-import { deleteNote } from '../redux/actions/dataActions';
+import { deleteNote } from '../../redux/actions/dataActions';
 
 const styles = theme => ({
   ...theme.spread,
