+++ src/components/profile/EditDetails.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import withStyles from '@material-ui/core/styles/withStyles';
-import AaButton from '../util/AaButton';
+import AaButton from '../../util/AaButton';
 // mui
 import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
@@ -14,7 +14,7 @@ import EditIcon from '@material-ui/icons/Edit';
 
 // redux
 import { connect } from 'react-redux';
-import { editUserDetails } from '../redux/actions/userActions';
+import { editUserDetails } from '../../redux/actions/userActions';
 
 const styles = theme => ({
   ...theme.spread,
